chore: split app
