+++ backend/backend/settings/localsync/common.py
@@ -1,0 +1,15 @@
+from ..deps.ledger import *
+from ..deps.orchestrator import *
+from ..dev import *
+
+INSTALLED_APPS = [
+    "django.contrib.auth",
+    "django.contrib.contenttypes",
+    "django_celery_results",
+    "rest_framework",
+    "localrep",
+    "localsync",
+    "substrapp",
+    "users",
+    "organization",
+]
