+++ backend/backend/settings/localsync/dev.py
@@ -1,0 +1,1 @@
+from .common import *
