+++ backend/backend/settings/localsync/prod.py
@@ -1,0 +1,1 @@
+from .common import *
