+++ skaffold.yaml
@@ -31,6 +31,7 @@ deploy:
           worker.image: substra/substra-backend
           schedulerWorker.image: substra/substra-backend
           events.image: substra/substra-backend
+          localsync.image: substra/substra-backend
 
       - name: backend-org-2
         chartPath: charts/substra-backend
@@ -47,6 +48,7 @@ deploy:
           worker.image: substra/substra-backend
           schedulerWorker.image: substra/substra-backend
           events.image: substra/substra-backend
+          localsync.image: substra/substra-backend
         skipBuildDependencies: true # deps already built by other org-1 release
   kubectl:
     manifests:
